+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/analysis/typing/reasoner/LocalTypeReasoner.java
@@ -32,6 +32,7 @@ public class LocalTypeReasoner extends AstWalker {
     private final LocalTypeReasonerState state;
     private final AssignmentHandler assignmentHandler;
     private final AtomHandler atomHandler;
+    private final ConditionalBodyHandler conditionalBodyHandler;
     private final ExpressionHandler expressionHandler;
     private final IdentifierHandler identifierHandler;
     private final InvocationHandler invocationHandler;
@@ -49,6 +50,7 @@ public class LocalTypeReasoner extends AstWalker {
 
         this.assignmentHandler = new AssignmentHandler(this.state);
         this.atomHandler = new AtomHandler(this.state);
+        this.conditionalBodyHandler = new ConditionalBodyHandler(this.state);
         this.expressionHandler = new ExpressionHandler(this.state);
         this.identifierHandler = new IdentifierHandler(this.state);
         this.invocationHandler = new InvocationHandler(this.state);
@@ -88,6 +90,11 @@ public class LocalTypeReasoner extends AstWalker {
     @Override
     protected void walkPostBody(final AstNode node) {
         this.expressionHandler.handleBody(node);
+    }
+
+    @Override
+    protected void walkPostConditionalExpression(final AstNode node) {
+        this.conditionalBodyHandler.handleConditionalExpression(node);
     }
 
     @Override